+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/ModelAndView.java
@@ -1,6 +1,6 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.mvc;
 
-import com.interface21.webmvc.servlet.view.View;
+import com.interface21.webmvc.servlet.mvc.view.View;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Collections;
