+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/resolver/HandlerResolver.java
@@ -1,10 +1,10 @@
 package com.interface21.webmvc.servlet.mvc.handler.resolver;
 
-import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.mvc.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.controller.Controller;
 import com.interface21.webmvc.servlet.mvc.handler.mapping.annotation.HandlerExecution;
-import com.interface21.webmvc.servlet.view.View;
-import com.interface21.webmvc.servlet.view.resolver.ViewResolver;
+import com.interface21.webmvc.servlet.mvc.view.View;
+import com.interface21.webmvc.servlet.mvc.view.resolver.ViewResolver;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
